typings: Split the DeviceState namespace types to a different file
Change-type: patch
